LKHWebsite/requirements.txt
django-bootstrap-navbar == 0.0.22
django-environ==0.4.5
uWSGI>=2.0.18,<2.1